feat: 발신 히스토리뷰 구현 및 이미지 로딩 적용
