+++ LovelyCrane/Manager/LetterListsManager.swift
@@ -25,10 +25,21 @@ final class LetterListsManager {
     }
     
     var sentLettersGroupedByDate: [Date: [LetterModel]] {
-        return Dictionary(grouping: sentLetters, by: { Calendar.current.startOfDay(for: $0.date) })
+        let groupByDate = Dictionary(grouping: sentLetters, by: { Calendar.current.startOfDay(for: $0.date) })
+        return sortGroupedLettersByDateDescending(groupByDate)
     }
     
     var notSentLettersGroupedByDate: [Date: [LetterModel]] {
-        return Dictionary(grouping: notSentLetters, by: { Calendar.current.startOfDay(for: $0.date) })
+        let groupByDate = Dictionary(grouping: notSentLetters, by: { Calendar.current.startOfDay(for: $0.date) })
+        return sortGroupedLettersByDateDescending(groupByDate)
+    }
+    
+    private func sortGroupedLettersByDateDescending(_ groupedLetters: [Date: [LetterModel]]) -> [Date: [LetterModel]] {
+        var sortedGroupedLetters: [Date: [LetterModel]] = [:]
+        for (date, letters) in groupedLetters {
+            let sortedLetters = letters.sorted(by: { $0.date > $1.date })
+            sortedGroupedLetters[date] = sortedLetters
+        }
+        return sortedGroupedLetters
     }
 }
