+++ LovelyCrane/View/WriteHistory/WriteHistoryView.swift
@@ -1,81 +1,133 @@
 //
-//  WriteHistoryView.swift
-//  LoveCrane
+//  WriteHistoryViewDev.swift
+//  LovelyCrane
 //
-//  Created by 235 on 2023/07/13.
+//  Created by Toughie on 2023/07/27.
 //
 
 import SwiftUI
 
-
 struct WriteHistoryView: View {
-    @State var isCellTapped = false
+    @Environment(\.dismiss) var dismiss
     
+    var letterCount: Int {
+        LetterListsManager.shared.isByMeLetters.count
+    }
+    var sentLetters: [Date : [LetterModel]] {
+        LetterListsManager.shared.sentLettersGroupedByDate
+    }
+    
+    var notSentLetters: [Date : [LetterModel]] {
+        LetterListsManager.shared.notSentLettersGroupedByDate
+    }
+
     var body: some View {
-        ZStack {
-            NavigationLink("", destination: DetailView(), isActive: $isCellTapped)
-            NavigationView {
-                ScrollView {
+        ZStack(alignment: .topLeading) {
+            Color(.backGround).ignoresSafeArea()
+            
+            //MARK: 최상단 VStack
+            VStack(alignment:.leading) {
+                //MARK: 헤더
+                HStack(alignment: .center) {
                     VStack(alignment: .leading) {
-                        LazyVStack {
-                            HistoryListCell(isRead: true)
-                                .onTapGesture {
-                                    isCellTapped.toggle()
+                        Text("나의 쪽지")
+                            .foregroundColor(.gray)
+                        HStack(alignment: .firstTextBaseline) {
+                            Text("\(letterCount)")
+                                .font(.largeTitle)
+                                .foregroundColor(.lightPink)
+                            Text("마리")
+                                .font(.title2)
+                                .foregroundColor(.white)
+                        }
+                    }
+                    Spacer()
+                    Image("WriteViewCranes")
+                }
+                .padding(.horizontal, 15)
+                .padding(.leading, 5)
+                
+                ScrollView {
+                    
+                    ForEach(Array(notSentLetters.keys.sorted(by: >)), id: \.self) { date in
+                        //MARK: 안 보낸 편지들
+                        LazyVStack(alignment: .leading) {
+                            if let notSentLettersGroup = notSentLetters[date] {
+//                                HStack {
+//                                   Image("historyCrane")
+//
+//                                    Text("\(notSentLettersGroup.count)마리의 종이학을 발송했어요 :) -\(formatDate(date))")
+//                                        .font(.caption)
+//                                        .foregroundColor(.white)
+//                                }
+//                                .padding(.leading, 20)
+
+                                
+                                ForEach(notSentLettersGroup, id: \.self) { letter in
+                                    WriteHistoryCell(letter: letter)
+                                }
+                            } else {
+                                Text("바인딩 실패패패패패퍂")
                             }
                         }
-                        header()
-                        LazyVStack {
-                            HistoryListCell(isRead: false)
-                                .onTapGesture {
-                                    isCellTapped.toggle()
+                        .padding(.bottom)
+                    }
+                    
+                    // MARK: 보낸 편지들
+                    ForEach(Array(sentLetters.keys.sorted(by: >)), id: \.self) { date in
+                        //MARK: 안 보낸 편지들
+                        LazyVStack(alignment: .leading) {
+                            if let sentLettersGroup = sentLetters[date] {
+                                HStack {
+                                   Image("historyCrane")
+
+                                    Text("\(sentLettersGroup.count)마리의 종이학을 발송했어요 :)")
+                                        .font(.caption)
+                                        .foregroundColor(.white)
+                                    
+                                    Text("-" + formatDate(date))
+                                        .font(.caption)
+                                        .foregroundColor(Color.tertiaryLabel)
+                                }
+                                .padding(.leading, 20)
+
+                                
+                                ForEach(sentLettersGroup, id: \.self) { letter in
+                                    WriteHistoryCell(letter: letter)
+                                }
+                            } else {
+                                Text("바인딩 실패패패패패퍂")
                             }
                         }
-                        Spacer()
+                        .padding(.bottom)
                     }
                 }
-                .padding()
-                .background {
-                    BackGroundView()
-                }
-            }.navigationTitle("")
-                .toolbar {
-                    ToolbarItem(placement: .navigationBarLeading) {
-                        toolBarTitle()
+            }
+            .padding(5)
+        }
+        .navigationBarBackButtonHidden(true)
+        .toolbar {
+            ToolbarItem(placement: .navigationBarLeading) {
+                Image(systemName: "chevron.left")
+                    .foregroundColor(.gray)
+                    .onTapGesture {
+                        dismiss()
                     }
-                }
+            }
         }
     }
-    
-    private func header() -> some View {
-        HStack {
-            Image(Assets.sendBottle)
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-            Text("97마리의 종이학을 발송했어요")
-                .foregroundColor(.white)
-            //텍스트랑 동일한 사이즈로 설정해줘야함.
-        }.frame(height: 18)
-            .padding(.vertical)
-    }
-    
-    private func toolBarTitle() -> some View {
-        HStack {
-            Text("9127개")
-                .font(.system(size: 26))
-                .foregroundColor(.white)
-            Image(Assets.conceptCrane)
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(height: 26)
-        }
-    }
-    
 }
 
-
-
-struct WriteView_Preview: PreviewProvider {
+struct WriteHistoryViewDev_Previews: PreviewProvider {
     static var previews: some View {
         WriteHistoryView()
+    }
+}
+
+extension WriteHistoryView {
+    private func formatDate(_ date: Date) -> String {
+    let dateFormatter = DateFormatter()
+    dateFormatter.dateFormat = "yyyy.MM.dd"
+    return dateFormatter.string(from: date)
     }
 }
